Big Update
Bot should better handle 2 users in chat now, prioritizing the game over just chatting
This update puts more emphasis on games. If 2 users are just chatting, 1 is playing league and one is playing minecraft. It will choose league or minecraft as the name. Even though more members are just chatting, I personally want it to focus hard on the gaming aspect of the users